Make float casts less aggressive
- Function arguments $f12, $f14 could be either float or double
- Conversion instructions (like `cvt.s.w`) do not need to coerce their
inputs to the expected type. *Try* to unify the types, but if it
fails, ignore it.

+++ tests/end_to_end/doubles/irix-g-mips1-out.c
@@ -1,9 +1,9 @@
-f64 test(f32 arg0, s32 arg2, f64 arg4) {
+f64 test(f64 arg0, s32 arg2, f64 arg4) {
     f64 sp8;
     f32 sp4;
     f64 sp0;
 
-    sp8 = ((f64) arg2 * (f64) arg0) + ((f64) arg0 / arg4);
+    sp8 = ((f64) arg2 * arg0) + (arg0 / arg4);
     sp8 -= 7.0;
     if ((sp8 < arg4) || (sp8 == arg4) || (sp8 > 9.0)) {
         sp0 = 2.3125f;
