+++ tests/end_to_end/int-float-doublelit/irix-g-out.c
@@ -5,5 +5,5 @@ f32 test(f32 arg0, s32 arg1, f32 arg2, s32 arg3) {
     arg2 = (f32) ((f64) arg2 + 0.0);
     arg2 += 0.0f;
     arg2 = (f32) ((f64) arg2 + D_400180);
-    return (f32) (u32) (arg3 + 3) + arg2;
+    return (f32) (arg3 + 3) + arg2;
 }
