+++ tests/end_to_end/int-float/irix-g-out.c
@@ -2,5 +2,5 @@ f32 test(f32 arg0, s32 arg1, f32 arg2, s32 arg3) {
     D_410130 = (s32) arg0;
     D_410134 = (f32) arg1;
     arg2 += 5.0f;
-    return (f32) (u32) (arg3 + 3) + arg2;
+    return (f32) (arg3 + 3) + arg2;
 }
